On the 'cmake' branch: Generate *.def files for libraries to support using shared DLLs on Windows.
When compiling shared library for Windows, it is required to generate
file with a list of what the DLL export from itself and let the compiler
create a *.lib file based on this list.

For extracting the library exports, there is already a python script,
named extractor.py. This commit adds a simple CMake wrapper for this
script for using it simply. Also, there is a field in the build.conf
file with the list of the headers to export from the library. This list
of headers in passed to the target_exports CMake function and then directly
to the extractor.py script to generate the *.def file.

After that, the generated *.def files are being passed to CMake  target via
target_sources() command. This command is invocated from the target_exports
to make the usage simpler.

In addition, these *.def and *.lib are required only for Windows, because
on Linux there is a different system with shared libraries and definitions
are generated automatically.

About the function name: the decision to name the function for setting up
the exports of the target was done to keep the CMake command naming style.
In CMake, there are commands named like target_include_directories,
target_sources, target_link_libraries, so the function can be named in this
style as target_exports. Now the generated CMakeLists.txt file looks very
nice due to a list of simple commands and file sources. Also this naming
abstraction allows to move the `if (WIN32)` check to this function.

* build/cmake/extractor.cmake: New file.
  (target_exports): New function.

* build/generator/gen_cmake.py
  (cmake_target): Add the msvc_export field.
  (Generator.write): Pass the msvc_export files to the ezt template.

* build/generator/templates/CMakeLists.txt.ezt
  (): Include the `extractor.cmake` file to import the run_extractor function.
  (targets loop): Generate run_extractor invocation for each library.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/cmake@1918852 13f79535-47bb-0310-9956-ffa450edef68