On the 'cmake' branch: Follow-up to r1918852: Fix argument parsing of the target_exports function.
After more experiments with the extractor, I found that arguments to the
target_exports function was passing incorrectly; the list of the
remaining exports were not working.

This problem can be fixed by removing the argument variable from the function
definition and using built-in CMake function `ARGN` for the remaining list.
This variable stores all remaining arguments passed to the function, except
declared arguments, in this case `target_name`. There is a similar to `ARGN`
variable, `ARGV`, but it is a little bit differ, because it contains all
variable passed to the function, including all declared in function definition
arguments.

* build/cmake/extractor.cmake
  (target_exports): CMake magic to pass the arguments to the extractor.py
   correctly.

git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/cmake@1918853 13f79535-47bb-0310-9956-ffa450edef68