On the 'cmake' branch: Support per-target build target customizations.
It is required to build some libraries in the static configuration even if
targeting shared libraries. There is a field in the build.conf file named
`msvc_static` for this.

In the CMake generator it is implemented by adding a build_type variable to the
cmake_target and later passing it to the ezt template where it will be added
to the library declaration. Currently it supports only the msvc_static field
from the build.conf file, but in feature, will be added additional options
for configuring the build type for the specific targets (for example, to build
FS modules statically). 

* build/generator/gen_cmake.py
  (cmake_target): Add the build_type field.

  (Generator.write): Initialize the build_type variable and pass it to the
   ezt template as explained above.

* build/generator/templates/CMakeLists.txt.ezt
  (): Use build_type variable for the library targets.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/cmake@1918859 13f79535-47bb-0310-9956-ffa450edef68