On the 'cmake' branch: Add BUILD_SHARED_LIBS option.
Now, when required preparations for shared build are done, this option can
be added and enabled.

Little regression: when using shared libraries and compiling test-suite,
some tests would not build, because they requires static libraries. The check
on it would be added later.

By default this option is enabled (set to ON).

This option is taken from the official documentation of CMake [1].

* build/cmake/options.cmake
  (): Add BUILD_SHARED_LIBS option.

[1] https://cmake.org/cmake/help/latest/variable/BUILD_SHARED_LIBS.html

git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/cmake@1918860 13f79535-47bb-0310-9956-ffa450edef68