On the 'cmake' branch: Disable target if msvc_force_static is true and building shared libraries.
This fixes TODO from r1918860 [1] and now it is possible to build shared libraries
and tests at the same time, but these tests would not be built.

[1]: [[[
Little regression: when using shared libraries and compiling test-suite,
some tests would not build, because they requires static libraries. The check
on it would be added later.
]]]

* build/generator/gen_cmake.py
  (Generator.write): Add `NOT BUILD_SHARED_LIBS` string to the enable_condition
   if msvc_force_static.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/cmake@1918861 13f79535-47bb-0310-9956-ffa450edef68