On the 'cmake' branch: Support build type customization for modules.
Add SVN_BUILD_SHARED_FS and SVN_BUILD_SHARED_RA options that allow to
customize the build type of the FS and RA modules.

Now, it is impossible to build RA modules statically, because they doesn't
support it.

The SVN_BUILD_SHARED_FS option defaults to ON when BUILD_SHARED_LIBS and to
OFF in other case. This is similar to simply defaulting this option to the
value of BUILD_SHARED_LIBS option. The SVN_BUILD_SHARED_RA option always
defaults to OFF, because it doesn't support shared build. In addition, adding
a warning when SVN_BUILD_SHARED_RA is set to true.

* build/cmake/options.cmake
  (): Add the options, explained above.
  (): Initialize SVN_FS_BUILD_TYPE and SVN_RA_BUILD_TYPE CMake variable, based
   on the options explained above.
  (): Write FATAL_ERROR if SVN_BUILD_SHARED_RA is set to true. 

* build/generator/gen_cmake.py
  (Generator.write): Set build_type to SVN_**_BUILD_TYPE for FS and RA modules.
  (Generator.write): Minor rework in sequence of build_type initialization.

git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/cmake@1918862 13f79535-47bb-0310-9956-ffa450edef68