On the 'cmake' branch: Factor out read_version function.
The same CMake magic to extract version from the includes is used multiple
times for different packages, so this code could become a function.

Note: the function is not going to be used from APR and APRUtil modules
(which also have a similar code), because they should be self-contained
and don't depend on any external functions by design, so the magic is
copied to these modules.

Note 2: Version of SQLite is recorded in non-standard form and the
function cannot read its version.

* CMakeLists.txt
  (read_version): New function with the CMake and regex magic.
  (Subversion, LZ4, and UTF8PROC versions): Replace CMake magic with
   read_version function invocation.

git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/cmake@1918991 13f79535-47bb-0310-9956-ffa450edef68