On the 'cmake' branch: Filter error codes to prevent warning.
When using `./gen-make -t cmake` command on Windows, there was a warning
related to error codes generation. The simplest solution is to filter-out
some error codes as it is done in the gen-win generator.

This is not the best solution, but there are no great alternative. In addition,
the file with the error codes would be empty in the tar-ball, so it's okay for
this code to contain checks for the platform.

Note: I decided to use 'sys.platform' constant instead of 'os.name' (which was
in the original patch [1]) for the code to be more like the the code from the
errno.pyi file. The errno.pyi file is part of the Python standard library and
it contains a list of error codes and it is used by the gen_base.py module.

Suggested by Jun Omae on dev@ [1].

* build/generator/gen_cmake.py
  (imports): Import 'sys' module for sys.platform constant.

  (errno_filter): New function, that filters the error code. Override this
   function only when platform is Windows.

[1] https://lists.apache.org/thread/6y0v5ozn2wmrp4vhxdf5h647rg9h5vmt

git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/cmake@1919039 13f79535-47bb-0310-9956-ffa450edef68