On the 'cmake' branch: Add SVN_DEBUG option and setup a definition based on it.
The SVN_DEBUG option could enable some developer features and log additional
information. Now, by passing `-DSVN_DEBUG=ON` argument to CMake, it would be
enabled. If this option is enabled, the define would be added for both, Debug
and Release configurations. The option is disabled by default.

SVN_DEBUG and just Debug are not the same things!!

* CMakeLists.txt
  (options): Add SVN_DEBUG option to cache and setup the same compile
   definition based on it.

git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/cmake@1919045 13f79535-47bb-0310-9956-ffa450edef68