On the 'cmake' branch: Use fake targets instead of variables for linking against entire list of FS and RA modules instead of variables.
In CMake, there is a possibility to create a target and without compiling
it. This kind of targets is called `interface targets' and declared with
the INTERFACE keyword.

Previously, for the module targets the system with adding them to CMake
variables (SVN_RA_MODULES and SVN_FS_MODULES) was used. However with the
interface targets it is possible to create a library with the list of the
modules and then simply link against  them by using the
`target_link_libraries(<target_name> [fs|ra]-libs)` command. The additional
feature of them is to declare libraries in any order.

In this commit, the libraries for the modules will be introduced and the FS
and RA modules will be linked into them as INTERFACE. Additionally, the
libraries which use the modules, will link against `**-libs` targets instead
of the lists of the modules.

* build/generator/gen_cmake.py
  (Generator.write): Use **-libs instead of SVN_**_MODULES for the FS and RA
   modules.

* build/generator/templates/targets.cmake.ezt
  (lib targets): Link the modules into **-libs targets as INTERFACE, remove
   previous code for appending the name into list.

* CMakeLists.txt
  (): Create empty ra-libs and fs-libs interface targets.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/cmake@1919103 13f79535-47bb-0310-9956-ffa450edef68