On the 'cmake' branch: gen_cmake.py: Remove the cmake_target class; use _ezt_data instead.
Till the cmake_target class doesn't do any significant work, it only now
complicates adding new features and fields to the generator. There were
already the _ezt_data class from the most initial version of the generator
that does a bit of magic in its constructor and creates an object with any
fields. It is already used for the 'data' of the generator and now, in this
commit, it will be used for the 'targets' list.

* build/generator/gen_cmake.py
  (cmake_target): Remove class.
  (Generator.write): Use _ezt_data instead of cmake_target for the
   new_target; add some transformations from cmake_target.

git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/cmake@1919217 13f79535-47bb-0310-9956-ffa450edef68