On the 'cmake' branch: Build static libraries without 'lib' prefix and fix bug with double 'lib' in name on Linux as implicit change.
Changes brief
-------------

- Skip 'lib' from the start of output name.

- Setup CMAKE_*_LIBRARY_PREFIX variables to add 'lib' to the start of shared
  library and implib names and to keep static libraries with the original
  name. 

How does it work?
-----------------

Firstly, when the output name is renamed to not contain 'lib' at the start, it
will build into just svn_client-1.lib, but since we set
CMAKE_SHARED_LIBRARY_PREFIX and CMAKE_IMPORT_LIBRARY_PREFIX to 'lib', it will
be prepended to the final output name automatically by CMake in shared
configuration, so it will be as it was before ('libsvn_client-1.lib'). Also
after this commit, the name will be left unchanged in static configuration
because CMAKE_STATIC_LIBRARY_PREFIX is set to empty. This repeats the behavior
of gen-win.

The another thing it fixes implicitly is the library names on Linux. On Linux
the CMAKE_SHARED_LIBRARY_PREFIX and CMAKE_IMPORT_LIBRARY_PREFIX variables are
automatically set to be 'lib', so before the OUTPUT_NAME has changed, the
libraries was compiling into files with double 'lib' in the name
(liblibsvn_client-1.so), which is incorrect. After this commit, their
behavior should be same as on Windows.

Also do a little refactoring of the abstraction of setting the name: the
output name is now passed from the generator (gen-cmake.py) to the template
using a field. This field contains a name without 'lib' prefix and also
'-1' postfix for major version. The template now only adds the content of
targets.output_name OUTPUT_NAME without any conversions.

Changes summary
---------------

* build/generator/gen_cmake.py
  (get_output_name): New function.
  (Generator.write): Add output_name field to the target.

* build/generator/templates/targets.cmake.ezt
  (output name): Simply pass output_name here instead of target name with
   '-1' added to the end.

* CMakeLists.txt
  (CMAKE_*_LIBRARY_PREFIX): Setup these vars and add some comments explaining
   them.

git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/cmake@1919320 13f79535-47bb-0310-9956-ffa450edef68