On the 'cmake' branch: gen_cmake.py: Refactoring in enable condition.
Factor-out a function that retrieves a list of enable conditions (condition
of when to enable the target). This prevents the code duplication and fixes
the abstraction of condition. The condition is retrieving for the target
directly and also for its dependencies, because the target cannot be built
if its dependency is unavailable, for example, fsfs-tests needs libsvn_fsfs
to compile successfully. When the function is factored-out, it could be
invoked for the target and for its dependencies. For example, we add
SVN_BUILD_TESTS for fsfs-tests and then, when handling libsvn_fsfs, 
SVN_ENABLE_FS_FS will be added to the enable condition list.

* build/generator/gen_cmake.py
  (get_target_conditions): New function.

  (Generator.write): Cut-off building of enable condition since it is moved
   into get_target_conditions() and append to the list return value of this
   function.

git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/cmake@1919347 13f79535-47bb-0310-9956-ffa450edef68