On the 'cmake' branch: Add checks for some functions and header files.
Use CheckIncludeFiles and CheckSymbolExists built-in CMake modules to
do checks of header files and functions for existing. These modules will try
to compile a test program to check if the header file or other symbol exists
on the system compiling to.

After the checks are done, setup compile definitions HAVE_* for that symbols
or headers using add_compile_definitions() command. In feature, these
defines could be moved into svn_private_config.h, to be similar as in
autoconf.

The list of checks is not yet full and needs to be expanded.

When implementing new checks in CMake, it is useful to check at what is
done in autoconf by looking at configure.ac file, configure logs, and the
produced svn_private_config.h file.

* CMakeLists.txt
  (checks): Add checks and include modules for them.

git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/cmake@1919380 13f79535-47bb-0310-9956-ffa450edef68