On the 'cmake' branch: Fix NULL device name for Unix.
On Windows, the NULL device is just a file with name 'nul' in any
directory, while on Unix, there is a file at '/dev/null' for this.
In this commit, an ifdef with the check for WIN32 will be added before
setting up the SVN_NULL_DEVICE_NAME definition to make it correct on
the current platform.

* subversion/svn_private_config.hc
  (SVN_NULL_DEVICE_NAME): Check for platform to setup the NULL device
   name correctly.

git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/cmake@1919409 13f79535-47bb-0310-9956-ffa450edef68