On the 'cmake' branch: Use private config for check results.
Write the result of the checks to the svn_private_config.h, to make it
more similar to autoconf, which use this file for that. Also it makes it
easier debug and possibly adjust the checks. To add the definitions correctly,
add a variable for them. Afterwards, when we should add a definition to
the file, we will append few strings with comment, definition, and newline
to the variable. At the end of configure this variable will be configured
into the private config. 

* CMakeLists.txt
  (PRIVATE_CONFIG_DEFINITIONS): Declare to empty string.
  (autocheck_include_files,
   autocheck_symbol_exists): Write the result into PRIVATE_CONFIG_DEFINITIONS
   instead of using add_compile_definitions() command. 
  (private config): Move it down for the PRIVATE_CONFIG_DEFINITIONS variable
   to be prepared for the configure.

* subversion/svn_private_config.hc
  (PRIVATE_CONFIG_DEFINITIONS): Insert it into the file; don't leave any
   newlines to make it prettier, because PRIVATE_CONFIG_DEFINITIONS already
   contains them.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/cmake@1919443 13f79535-47bb-0310-9956-ffa450edef68