py-tests: Correctly search for tools in install layout.
There is already a support for the install layout in the Python test suite,
but some tools wasn't supporting it yet. This might cause an error like
entriesdump not found, so tests fail with this binaries layout. Additionally
there is also a support of some binaries in install layout, but others still
not working.

Install layout means that the binaries are installed in a single directory.
This kind of layout will be also used for build with CMake, while with
autoconf and gen-win build systems they are installed each in their directory.

The current fix will add the checks for the --tools-bin --bin argument for
all binaries in the main of the Python test suite and later using the binaries
as install layout if so.

[in subversion/tests/cmdline/svntest]

* main.py
  (execute_tests, entry): Making some variables to be global.
  (execute_tests, binary paths fixup): Handle all binaries for install layout
  inside the checks for --bin and --tools-bin arguments.


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1919489 13f79535-47bb-0310-9956-ffa450edef68