Support finding svnserve from install layout for ra-test.
This is implemented by adding a function for finding svnserve binary
and checking both files for existing. Firstly, check for the svnserve,
located in the ../../svnserve/svnserve[.exe] and finally will try just
svnserve[.exe].

Adding a preprocessor definition for executable extension on the current
platform as a little refactoring, required for this change.

Additionally, change the error message if the binary not found a bit,
because it needs to write both possible paths of svnserve binary.

This will be needed for running the test correctly with CMake due to the
different build layout.

[in subversion/tests]

* libsvn_ra/ra-test.c
  (EXE_EXTENSION): Define it.
  (find_svnserve): New function that finds svnserve binary.
  (open_tunnel): Use find_svnserve() function for svnserve binary and remove
   the previous code for finding it.


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1919505 13f79535-47bb-0310-9956-ffa450edef68