On the 'cmake' branch: Add Python tests using the add_test() command.
We are currently, using built-in CMake tool for testing, CTest. In this
commit, the Python tests will be also added into it, as addition to the
C-Tests. Also, add finding of Python executable, which is required for
running the Python test-suite. These tests will be added only if the
SVN_BUILD_TESTS enabled, at the same time with c-tests. For retrieving
the tests, we will simply enumerate all *.py files in the
subversion/tests/cmdline directory. Afterwards, we will use the
run_tests.py script for running the tests and streaming the output.

* CMakeLists.txt
  (tests): Use add_test() command for each python test file to run them.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/cmake@1919625 13f79535-47bb-0310-9956-ffa450edef68