Add check for uname having before retrieving it in the linux_release_name function.
We have a wrapper over uname(); release_name_from_uname. This function would
be exist only if the HAVE_UNAME definition is true. However, the
linux_release_name function exists on every Linux system. So, this function
should work without uname. To support it, add a check for HAVE_UNAME inside
the linux_release_name function and request uname only if so.

* subversion/libsvn_subr/sysinfo.c
  (linux_release_name): Retrieve uname only when we have it.


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1919659 13f79535-47bb-0310-9956-ffa450edef68