On the 'cmake' branch: Reverting the change from r1918954 to use Module for finding Expat instead of config.
There is a problem when using CONFIG mode for expat library on the OpenBSD
system because the config file doesn't exist on this system. Using of module
would fix it.

The problem is reported and discussed in the `Building with CMake on Linux'
thread on dev@; the discussion is archived and available at [1].

* CMakeLists.txt
  (expat finding): Remove CONFIG keyword from find_package invocation to use
   Module for it and adjust the name used for expat since they are different
   from the names used in CONFIG mode.

Found by: Stefan Sperling
Suggested by: Daniel Sahlberg

[1] https://lists.apache.org/thread/1mpqrt98owsxoj13rmqlo1m8r0ll9cxv


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/cmake@1919660 13f79535-47bb-0310-9956-ffa450edef68