On the 'cmake' branch: Support compilation of Perl SWIG bindings.
For finding the Perl dev-kit, the FindPerlLibs.cmake built-in CMake module
will be utilized. This module doesn't actually create a target, so we will
do it manually and will just setup the external-perl target properties with
values from FindPerlLibs. The targets for the Perl SWIG bindings are already
created by gen-cmake while other Perl-specific options will be handled by the
UseSWIG module.

* CMakeLists.txt
  (swig.perl): Create the external-perl target for the Perl bindings.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/cmake@1919799 13f79535-47bb-0310-9956-ffa450edef68