On the 'cmake' branch: Fixup Perl flags list before using them.
We are utilizing the additional Perl compile options from the FindPerlLibs
module. However, they are incorrectly separated (they should be represented
as a CMake array, while they are recorded just as a plain string).

CMake provides an out-of-box function that could convert one format to another,
so we could just invoke it to fixup the compile options before usage.

This fix will resolve an issue when compiling the Perl SWIG bindings on Linux
with CMake. The problem was reported and discussed on the dev@ list. The thread
is publicly archived at [1].

* CMakeLists.txt
  (swig.perl): Use the separate_arguments() command to fixup compile options.

[1] https://lists.apache.org/thread/g7w5z12l9dw41lhc25zyqnvr298mcw95


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/cmake@1919813 13f79535-47bb-0310-9956-ffa450edef68