On the 'cmake' branch: fix 'SQLiteAmalgamation or not' problem.
There were an explicit CMake option that tells CMake which SQLite to use,
installed or the amalgamation. It was set to use the amalgamation by default
which is actually uncommon.

To resolve the issue, firstly the default is going to be changed to 'ON'.
Afterwards, if the option has not been changed (is still set to ON), we will
also try to find the amalgamation instead, and write a warning of implicitly
using SQLite amalgamation.

This follows that the user will use the installed library if they didn't
change this option, will use amalgamation if is is specified explicitly,
and also will use the amalgamation if the library wasn't found.

* CMakeLists.txt
  (options): Change default of the SVN_SQLITE_USE_AMALGAMATION option to OFF.
  (sqlite): Try SQLite amalgamation even if SVN_SQLITE_USE_AMALGAMATION tells
   not to use it.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/cmake@1920296 13f79535-47bb-0310-9956-ffa450edef68