Followup to r1919535.
err was already declared earlier in the same function. In addition the second
declaration was not on the top of the block, violating C90.
This was caught compiling with --enable-maintainer-mode.

We can reuse the existing pointer, since it is only used when an error occurs
and in that case the function immediately return.

* subversion/svnlook/svnlook.c
  (sub_main, handling option l): Re-use existing err pointer

Found by: rinrab


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1920521 13f79535-47bb-0310-9956-ffa450edef68