GitHub Actions: Configure initial workflow.
This workflow has been already tested in my GitHub fork, and should work
correctly. It does the following:

- Building the project using CMake.
- Running on Windows and Linux platforms.
- Checks `install`, and shared/static configurations.
- Run the test-suite (C and Python).

Note: some tests are not yet passing with CMake, and will be fixed soon.

The idea of configuring the GitHub Actions has discussed on dev@ in the
'Using GitHub Actions for continuous integration' thread [1].

After this commit, the actions should be available on the GitHub mirror
of the project at [2].

[1] https://lists.apache.org/thread/42gpcjh06526kpkt6yx2d0dhv98qbl7d
[2] https://github.com/apache/subversion/actions

* .github/workflows/cmake.yml: New file with recursively created parents.


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1920766 13f79535-47bb-0310-9956-ffa450edef68