Fix compilation with CLang on Windows.
Before, we were invoking the svn_utf__win32_utf8_to_utf16() function with
non-constant pointer, passed into the 'result' argument. However, in the
svn_io__utf8_to_unicode_longpath function implementation, there is a
conversion to a constant pointer causing the following error when compiling
with CLang:

[[[
subversion\libsvn_subr\io.c(1902,42): error : cannot take the address of
an rvalue of type 'const WCHAR *' (aka 'const unsigned short *')
]]]

In this commit, this conversion will be removed in order to prevent the error,
how it is done in the other places, where the function is utilized.

There still will be a warning of implicit conversion, which has to be
fixed generally in this function.

* subversion/libsvn_subr/io.c
  (svn_io__utf8_to_unicode_longpath): Do not convert 'buffer' to 'const WCHAR *'
   before taking address of it.


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1921361 13f79535-47bb-0310-9956-ffa450edef68