cmake: Factor-out add_private_config_definition macro for adding definitions to the private config.
In the past, we had to append and format a string to the
PRIVATE_CONFIG_DEFINITIONS variable manually in order to create new
definitions in svn_private_config.h header. This commit introduces
a new function add_private_config_definition() for this, which will
build the string and append it to the file.

* CMakeLists.txt
  (add_private_config_definition): New macro.
  (): Use the add_private_config_definition function everywhere.


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1921379 13f79535-47bb-0310-9956-ffa450edef68