Factor-out definitions of the utility functions for manupaluting with merge_source_t into a private client.h header file, in order to use them in other source files as well.
The merge_source_t struct has already been defined in this file, while
merge_source_create() and merge_source_dup() function were implemented
directly in the merge.c file (merge implementation itself). Since the struct
itself is defined the header file, and is already utilized in the other file,
it is going to be nice to have these utilities there.

* subversion/libsvn_client/client.h
  (svn_client__merge_source_create,
   svn_client__merge_source_dup): Declare functions.
* subversion/libsvn_client/merge.c
  (svn_client__merge_source_create,
   svn_client__merge_source_dup): Renamed implementation to include the
   `svn_client__` preffix, remove static.
  (...everywhere...): Adjust usages of these functions.


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1921433 13f79535-47bb-0310-9956-ffa450edef68