Create 'xml-writer' branch to implement an XML writer.
XML writer is a set of functions for creating xml tags and writing them
directly to a general svn stream without temporary buffers.

This is my old work, but it's not ready for 1.15 trunk, so I decided to
commit it to a branch instead. I'll be very happy to receive any feedback.

It's mostly made to implement xpatch feature, since it would require a lot
of xml writes, but can be also used for anything in subversion, including
cmdline with `--xml` option, and maybe even ra-serf library.

git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/xml-writer@1925685 13f79535-47bb-0310-9956-ffa450edef68