branches/1.14.x/STATUS: Remove obsolete nomination
The r1928521, r1928524 group was nominated twice: originally without a
conflict resolution branch, and subsequently with.

The 1.14.x-r1928521 conflict resolution branch was voted and approved in
r1929899 and r1929558, so the original conflicting nomination is no longer
needed. Removing it will also stop the GitHub Actions backportbot's complaints
on this branch, such as this one:
https://lists.apache.org/thread/q154c36q23jq03cxt00n107w9244yt83


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/1.14.x@1929905 13f79535-47bb-0310-9956-ffa450edef68